+++ server/useCase/playerUseCase.ts
@@ -8,7 +8,7 @@ import { gameOver } from '$/service/gameOver';
 import { userIdParser } from '$/service/idParsers';
 import { isInDisplay } from '$/service/isInDisplay';
 import { minmax } from '$/service/minmax';
-import { posWithDirSpeTim } from '$/service/posWithDirSpeTim';
+import { posWithBulletModel } from '$/service/posWithBulletModel';
 import { randomUUID } from 'crypto';
 import { bulletUseCase } from './bulletUseCase';
 
@@ -51,59 +51,6 @@ export const playerUseCase = {
     await gameOver(player, newPlayer);
     await bulletsRepository.delete(bulletId);
   },
-  trackingBullets: async (player: PlayerModel) => {
-    const res = await enemiesRepository.findNotNull();
-    const lockOnEnemies = sortByDistance(player, res).filter(
-      (enemy) => enemy.pos.x > player.position.x
-    );
-    if (lockOnEnemies.length === 0) return;
-    Promise.all(
-      lockOnEnemies.map((enemy) => {
-        const diffX = enemy.pos.x - player.position.x;
-        const diffY = enemy.pos.y - player.position.y;
-        const normalization = 1 / Math.sqrt(enemy.squaredDistance);
-        const dir = {
-          x: diffX * normalization,
-          y: diffY * normalization,
-        };
-        bulletUseCase.createByPlayer(player, dir);
-      })
-    ).then((results) =>
-      results.forEach((result) => {
-        result;
-      })
-    );
-  },
-  manyBullets: async (player: PlayerModel): Promise<void> => {
-    const subShotYs = (numOfBullet: number) =>
-      [...Array(numOfBullet)].map((_, i) => [-1, 1].map((n) => i * n));
-    Promise.all(
-      subShotYs(3)
-        .flat()
-        .map((y: number) => {
-          const normalization = Math.sqrt(1 + y ** 2);
-          const dir = {
-            x: 1,
-            y: y * normalization,
-          };
-          return bulletUseCase.createByPlayer(player, dir);
-        })
-    );
-  },
-  barrier: async (player: PlayerModel) => {
-    const res = await bulletsRepository.findAllByEnemy();
-    const deletingBullets = res.filter((bullet) => {
-      const [x, y] = posWithDirSpeTim(bullet);
-      const BARRIER_WIDTH = 100;
-      return (player.position.x - x) ** 2 + (player.position.y - y) ** 2 < BARRIER_WIDTH;
-    });
-    Promise.all(deletingBullets.map((bullet) => bulletsRepository.delete(bullet.id))).then(
-      (results) =>
-        results.forEach((result) => {
-          result;
-        })
-    );
-  },
   create: async (userName: string): Promise<PlayerModel | null> => {
     const newPlayer: PlayerModel = {
       id: userIdParser.parse(randomUUID()),
@@ -129,5 +76,60 @@ export const playerUseCase = {
         position: { ...player.position, x: player.position.x - 1920 * displayNumber },
       }));
     return playerInDisplay;
+  },
+  item: {
+    trackingBullets: async (player: PlayerModel) => {
+      const res = await enemiesRepository.findNotNull();
+      const lockOnEnemies = sortByDistance(player, res).filter(
+        (enemy) => enemy.pos.x > player.position.x
+      );
+      if (lockOnEnemies.length === 0) return;
+      Promise.all(
+        lockOnEnemies.map((enemy) => {
+          const diffX = enemy.pos.x - player.position.x;
+          const diffY = enemy.pos.y - player.position.y;
+          const normalization = 1 / Math.sqrt(enemy.squaredDistance);
+          const dir = {
+            x: diffX * normalization,
+            y: diffY * normalization,
+          };
+          bulletUseCase.createByPlayer(player, dir);
+        })
+      ).then((results) =>
+        results.forEach((result) => {
+          result;
+        })
+      );
+    },
+    manyBullets: async (player: PlayerModel): Promise<void> => {
+      const subShotYs = (numOfBullet: number) =>
+        [...Array(numOfBullet)].map((_, i) => [-1, 1].map((n) => i * n));
+      Promise.all(
+        subShotYs(3)
+          .flat()
+          .map((y: number) => {
+            const normalization = Math.sqrt(1 + y ** 2);
+            const dir = {
+              x: 1,
+              y: y * normalization,
+            };
+            return bulletUseCase.createByPlayer(player, dir);
+          })
+      );
+    },
+    barrier: async (player: PlayerModel) => {
+      const res = await bulletsRepository.findAllByEnemy();
+      const deletingBullets = res.filter((bullet) => {
+        const [x, y] = posWithBulletModel(bullet);
+        const BARRIER_WIDTH = 100;
+        return (player.position.x - x) ** 2 + (player.position.y - y) ** 2 < BARRIER_WIDTH;
+      });
+      Promise.all(deletingBullets.map((bullet) => bulletsRepository.delete(bullet.id))).then(
+        (results) =>
+          results.forEach((result) => {
+            result;
+          })
+      );
+    },
   },
 };
